+++ pytest.ini
@@ -3,7 +3,7 @@ xvfb_width = 1280
 xvfb_height = 800
 # Xvfb settings for GUI tests; remove temporary exclusions and run all tests
 xvfb_colordepth = 24
-addopts = -m "not slow"
+addopts = --suite core
 markers =
     unit: unit tests
     integration: integration tests (MCP/LLM/CLI)
@@ -11,3 +11,8 @@ markers =
     slow: long-running tests
     real_llm: tests that call a real OpenRouter endpoint
     smoke: quick smoke tests for core functionality
+    core: fast default suite (unit tests, smoke tests, and other local-only checks)
+    service: integration scenarios against the local MCP/CLI stack (includes core suite)
+    gui_smoke: small GUI subset that exercises the main windows without the full matrix
+    gui_full: comprehensive GUI regression suite
+    quality: static analysis and translation coverage checks
