+++ tests/gui/test_main.py
@@ -7,7 +7,7 @@ from typing import ClassVar
 
 import pytest
 
-pytestmark = pytest.mark.gui
+pytestmark = [pytest.mark.gui, pytest.mark.gui_smoke]
 
 
 def test_main_runs(monkeypatch):
